test: fail the browser suites when Chrome will not launch
They caught the launch error and passed with a console.warn, so an
environment without a browser reported the checks as green — under Bun,
WebpackDevServer.longtest.js passed in CI while all three of its tests did
nothing. The launch now lives in test/helpers/launchChrome.js and throws.

Jest on Bun cannot load the ESM-only puppeteer-core at all, so the three
browser suites are excluded from test:bun / test:deno at the runner level
rather than skipped from inside.
